Migrate MachineDeployment controller to Kubebuilder v2
